[onert] Revisit CompilerOptions
This commit revisits CompilerOptions.
- Move class implementation to CompilerOptions.h/cpp
- Move option seting methods into CompilerOptions
  - Move enableToFp16 into CompilerOptions
  - Introduce forceInternalOptions method
- Move verboseOptions into CompilerOptions

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
