+++ runtime/onert/core/include/compiler/Compiler.h
@@ -22,6 +22,7 @@
 #ifndef __ONERT_COMPILER_COMPILE_H_
 #define __ONERT_COMPILER_COMPILE_H_
 
+#include "CompilerOptions.h"
 #include "ir/NNPkg.h"
 #include "exec/Executors.h"
 #include "util/TracingCtx.h"
@@ -31,39 +32,6 @@ namespace onert
 
 namespace compiler
 {
-
-struct ManualSchedulerOptions
-{
-public:
-  void setBackendMap(const std::string &str);
-
-public:
-  std::string backend_for_all;
-  std::unordered_map<ir::OpCode, std::string> opcode_to_backend;
-  std::unordered_map<ir::OperationIndex, std::string> index_to_backend;
-};
-
-class CompilerOptions
-{
-public:
-  // Set default values for CompilerOptions
-  // All these default values should not be fetched from Env, when we stop supporting Android NNAPI.
-  static std::unique_ptr<CompilerOptions> fromGlobalConfig();
-
-public:
-  // GENERAL OPTIONS
-  std::vector<std::string> backend_list;
-
-  // OPTIONS ONLY FOR DEBUGGING/PROFILING
-  std::string trace_filepath; //< File path to save trace records
-  int graph_dump_level;       //< Graph dump level, values between 0 and 2 are valid
-  std::string executor;       //< Executor name to use
-  ManualSchedulerOptions manual_scheduler_options; //< Options for ManualScheduler
-  bool he_scheduler;      //< HEScheduler if true, ManualScheduler otherwise
-  bool he_profiling_mode; //< Whether HEScheduler profiling mode ON/OFF
-  bool disable_compile;   //< Run with Interpreter if true, try compilation otherwise
-  bool fp16_enable;       //< Whether fp16 mode ON/OFF
-};
 
 struct CompilerArtifact
 {
@@ -104,11 +72,6 @@ public:
    * @return std::shared_ptr<CompilerArtifact> Executors as a result of compilation
    */
   std::shared_ptr<CompilerArtifact> compile(void);
-
-  /**
-   * @brief   Allow to compute float32 using float16 data type
-   */
-  void enableToFp16();
 
 private:
   void checkProfilerConditions();
